extract ActionsButton
